Got vndb sql test to compile, still need to debug.
I had to remove all the json.get<json::string_view_t>, since they depend
on a from_json function for string_view, and I'm not up to writing that
right now. Nothing I changed should effect the results of anything, it
just may make things a bit slower.

+++ C++/util/string_buf.h
@@ -411,7 +411,7 @@ gen_append_fn(append_hex, double, "%a");
               "format string was \"%s\".\n", fmt);
       return (*this);
     }
-    if(nbytes >= space_left){
+    if((size_t)nbytes >= space_left){
       buf.reserve(nbytes+1);
       nbytes = snprintf(buf.end(), nbytes + 1,
                         fmt,  args...);
